Shuffled the offworld gate psionic curses to their own class.
